// src/actions/types.tsx
export enum types {
    SET_IS_LOADING = 'SET_IS_LOADING',
    INIT_APP = 'INIT_APP',
    INIT_AUTHENTICATION = 'INIT_AUTHENTICATION',
    INIT_SESSION_AUTH = 'INIT_SESSION_AUTH',
    AUTH_SUCCESS = 'AUTH_SUCCESS',
    AUTH_FAILURE = 'AUTH_FAILURE',
    INIT_LOGOUT = 'INIT_LOGOUT',
    SAVE_FORM_DATA_INTO_STORE = 'SAVE_FORM_DATA_INTO_STORE',
    SET_IS_USER_LOGGED_IN = 'SET_IS_USER_LOGGED_IN',
    SET_USER_DATA = 'SET_USER_DATA',
    SET_USERS = 'SET_USERS',
    SET_ROOMS = 'SET_ROOMS',
    SET_CURRENT_ROOM = 'SET_CURRENT_ROOM',
    SET_ACTIVE_ROOM_USERS = 'SET_ACTIVE_ROOM_USERS',
    SET_MESSAGES = 'SET_MESSAGES',
    INIT_SEND_GENERAL_MESSAGE = 'INIT_SEND_GENERAL_MESSAGE',
    INIT_SEND_ROOM_MESSAGE = 'INIT_SEND_ROOML_MESSAGE',
    SET_SOCKET_CONNECTION_STATUS = 'SET_SOCKET_CONNECTION_STATUS',
    INIT_ROOM_ENTER = 'INIT_ROOM_ENTER',
    INIT_ROOM_LEAVE = 'INIT_ROOM_LEAVE',
    INIT_ROOM_ADMIN_LEAVE = 'INIT_ROOM_ADMIN_LEAVE',
    INIT_ROOM_CREATE = 'INIT_ROOM_CREATE',
    INIT_HANDLE_ROOM_CREATE = 'INIT_HANDLE_ROOM_CREATE',
    INIT_ROOM_ADMIN_CHANGE = 'INIT_ROOM_ADMIN_CHANGE',
    INIT_CHECK_ROOM_PASSWORD = 'INIT_CHECK_ROOM_PASSWORD',
    INIT_CHECK_ROOM_PASSWORD_FAILURE = 'INIT_CHECK_ROOM_PASSWORD_FAILURE',
    INIT_ROOM_ENTER_SUCCESS = 'INIT_ROOM_ENTER_SUCCESS',
    INIT_CHECK_INBOX = 'INIT_CHECK_INBOX',
    SET_INBOX_MESSAGES = 'SET_INBOX_MESSAGES',
    INIT_SEND_INBOX_MESSAGE = 'INIT_SEND_INBOX_MESSAGE',
    INIT_RECEIVE_INBOX_MESSAGE = 'INIT_RECEIVE_INBOX_MESSAGE',
    SET_INBOX_COUNT = 'SET_INBOX_COUNt',
    SET_DRAWING_POINT = 'SET_DRAWING_POINT',
    SET_BROADCASTED_DRAWING_POINT = 'SET_BROADCASTED_DRAWING_POINT',
    SET_BROADCASTED_DRAWING_POINTS_BULK = 'SET_BROADCASTED_DRAWING_POINTS_BULK',
    SET_NEW_DRAWING_POINTS_GROUP = 'SET_NEW_DRAWING_POINTS_GROUP',
    SET_NEW_BROADCASTED_DRAWING_POINTS_GROUP = 'SET_NEW_BROADCASTED_DRAWING_POINTS_GROUP',
    INIT_CLEAR_DRAWING_POINTS = 'INIT_CLEAR_DRAWING_POINTS',
    CLEAR_DRAWING_POINTS = 'CLEAR_DRAWING_POINTS',
    INIT_DRAWING_BROADCAST = 'INIT_DRAWING_BROADCAST',
    INIT_MOUSE_UP_BROADCAST = 'INIT_MOUSE_UP_BROADCAST',
    SET_DRAW_COUNT = 'SET_DRAW_COUNT',
    SET_CURRENT_DRAWING = 'SET_CURRENT_DRAWING',
    INIT_CANVAS_TO_IMAGE = 'INIT_CANVAS_TO_IMAGE',
    INIT_GET_IMAGES_FROM_SERVER = 'INIT_GET_IMAGES_FROM_SERVER',
    SET_USER_DRAWINGS = 'SET_USER_DRAWINGS',
    INIT_CREATE_NEW_DRAWING = 'INIT_CREATE_NEW_DRAWING',
    INIT_DRAWING_SELECT = 'INIT_DRAWING_SELECT',
    INIT_IN_ROOM_DRAWING_SELECT = 'INIT_IN_ROOM_DRAWING_SELECT',
    NETWORK_ERROR = 'NETWORK_ERROR'
};